Overhaul database access: type-safe queries and Postgres support
* Move queries to JOOQ
* Set database charset on MySQL and MariaDB
* Remove MySQL syntax compatibility when using HSQLDB
* #48 - Require PostgreSQL 12 due to use of generated columns
* Progress on #73
* Solve #103
* Require PostgreSQL 12

+++ bans-api/src/main/java/space/arim/libertybans/api/revoke/RevocationOrder.java
@@ -44,7 +44,7 @@ public interface RevocationOrder {
 	 * 
 	 * @return the ID of the punishment to be revoked if there is one
 	 */
-	Optional<Integer> getID();
+	Optional<Long> getID();
 
 	/**
 	 * Gets the punishment type of the punishment which will be revoked, or none if
